# contracts/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "AnalyzeRequest": {
        "properties": {
          "spec_text": {
            "maxLength": 50000,
            "minLength": 10,
            "title": "Spec Text",
            "type": "string"
          },
          "submittal_text": {
            "maxLength": 50000,
            "minLength": 10,
            "title": "Submittal Text",
            "type": "string"
          },
          "system_id": {
            "default": "CUSTOM",
            "maxLength": 50,
            "title": "System Id",
            "type": "string"
          }
        },
        "required": [
          "spec_text",
          "submittal_text"
        ],
        "title": "AnalyzeRequest",
        "type": "object"
      },
...
      "Body_v1_analyze_upload": {
        "properties": {
          "spec_file": {
            "contentMediaType": "application/octet-stream",
            "title": "Spec File",
            "type": "string"
          },
          "submittal_file": {
            "contentMediaType": "application/octet-stream",
            "title": "Submittal File",
            "type": "string"
          }
        },
        "required": [
          "spec_file",
          "submittal_file"
        ],
        "title": "Body_v1_analyze_upload",
        "type": "object"
      },
      "Body_v1_analyze_upload_stream": {
        "properties": {
          "spec_file": {
            "contentMediaType": "application/octet-stream",
            "title": "Spec File",
            "type": "string"
          },
          "submittal_file": {
            "contentMediaType": "application/octet-stream",
            "title": "Submittal File",
            "type": "string"
          }
        },
        "required": [
          "spec_file",
          "submittal_file"
        ],
        "title": "Body_v1_analyze_upload_stream",
        "type": "object"
      },
      "Body_v1_analyze_vision": {
        "properties": {
          "spec_file": {
            "contentMediaType": "application/octet-stream",
            "title": "Spec File",
            "type": "string"
          },
          "submittal_image": {
            "contentMediaType": "application/octet-stream",
            "title": "Submittal Image",
            "type": "string"
          }
        },
        "required": [
          "spec_file",
          "submittal_image"
        ],
        "title": "Body_v1_analyze_vision",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Spec-to-Site Deviation Sentinel for hyperscale data-centre EPC delivery",
    "title": "Pramaan — EPC Deviation Intelligence",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/analyze": {
      "post": {
        "operationId": "v1_analyze",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 Analyze",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/analyze/stream": {
      "post": {
        "operationId": "v1_analyze_stream",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 Analyze Stream",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/analyze/upload": {
      "post": {
        "operationId": "v1_analyze_upload",
        "parameters": [
          {
            "in": "query",
            "name": "system_id",
            "required": false,
            "schema": {
              "default": "CUSTOM",
              "title": "System Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_v1_analyze_upload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 Analyze Upload",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/analyze/upload/stream": {
      "post": {
        "operationId": "v1_analyze_upload_stream",
        "parameters": [
          {
            "in": "query",
            "name": "system_id",
            "required": false,
            "schema": {
              "default": "CUSTOM",
              "title": "System Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_v1_analyze_upload_stream"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 Analyze Upload Stream",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/analyze/vision": {
      "post": {
        "description": "Reconcile a text spec against a submittal supplied AS AN IMAGE — Gemini\nvision reads values straight from the picture (datasheet page, table, or\ndrawing). Gemini-only (no text-model failover); on any failure returns\nmode='vision-unavailable' with no findings. Gate off with PRAMAAN_VISION=0.\nCapability evidence: data/samples/real/VISION_RESULT.md.",
        "operationId": "v1_analyze_vision",
        "parameters": [
          {
            "in": "query",
            "name": "system_id",
            "required": false,
            "schema": {
              "default": "CUSTOM",
              "title": "System Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_v1_analyze_vision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 Analyze Vision",
        "tags": [
          "v1"
        ]
      }
    },
...
    "/api/v1/jobs/analyze": {
      "post": {
        "description": "Submit an analysis as a background job; returns immediately (202) with a\njob_id to poll. Same auth + analysis rate limit as /analyze.",
        "operationId": "v1_submit_analyze_job",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 Submit Analyze Job",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "description": "Lightweight status metadata for a job (no result body) — poll this.",
        "operationId": "v1_get_job",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 Get Job",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/result": {
      "get": {
        "description": "The full analysis result once the job is done; 202 while still running,\n404 for an unknown/expired id.",
        "operationId": "v1_get_job_result",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 Get Job Result",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/llm-check": {
      "get": {
        "description": "Make a real LLM call and report the actual outcome. Unlike /health\n(which only checks a key is present), this surfaces the true reason\nanalysis falls back — e.g. out of credit, bad model, bad key — without\never returning the key itself.\n\nReports the full provider failover chain: the configured providers in the\norder they will be tried (gemini → gateway → claude), which one last\nanswered, the last failover reason, and — when no provider is configured —\nthat the system is running on the deterministic rule-engine floor.\n\n`?deep=1` sends the *same reconcile-sized prompt the live demo sends* (a\ntiny probe can pass while demo-sized calls 429 on token-weighted quotas),\nbounded by the analyze timeout, and reports honest mode/latency/findings.\n`?probe_all=1` tests every configured provider individually (uses one call\nper provider — spend sparingly on free tiers).",
        "operationId": "v1_llm_check",
        "parameters": [
          {
            "in": "query",
            "name": "deep",
            "required": false,
            "schema": {
              "default": false,
              "title": "Deep",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "probe_all",
            "required": false,
            "schema": {
              "default": false,
              "title": "Probe All",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V1 Llm Check",
        "tags": [
          "v1"
        ]
      }
    },
...
  }
}